+++ src/main/java/com/example/LearningLog/repos/EntryRepo.java
@@ -6,7 +6,7 @@ import com.example.LearningLog.models.Entry;
 
 public interface EntryRepo extends JpaRepository<Entry, Long> {
 
-    Iterable<Entry> findByTopicId(Integer topic_id);
-    Iterable<Entry> findByTopicIdOrderByDateTime(Integer topic_id);
+    Iterable<Entry> findByTopicId(Long topic_id);
+    Iterable<Entry> findByTopicIdOrderByDateTime(Long topic_id);
     
 }
